feat: Routes Protected and testing edited

+++ src/transaction/transaction.service.ts
@@ -1,10 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateTransactionDto } from './dto/create-transaction.dto';
 import { UpdateTransactionDto } from './dto/update-transaction.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Transaction } from './entities/transaction.entity';
 import { FindQueryParams } from './dto/findParams';
+import { User } from '../auth/entities/user.entity';
 
 @Injectable()
 export class TransactionService {
@@ -13,18 +18,22 @@ export class TransactionService {
     private readonly trRepo: Repository<Transaction>,
   ) {}
 
-  async create(dto: CreateTransactionDto): Promise<Transaction> {
-    return await this.trRepo.save(dto);
+  async create(dto: CreateTransactionDto, user: User): Promise<Transaction> {
+    return await this.trRepo.save({ ...dto, user });
   }
 
   private baseQuery() {
     return this.trRepo.createQueryBuilder('t').orderBy('t.createdAt', 'DESC');
   }
 
-  async findAll(queryParams: FindQueryParams) {
+  async findAll(queryParams: FindQueryParams, user: User) {
     const { categoryName, amountFrom, amountTo, createdFrom, createdTo } =
       queryParams;
-    const query = this.baseQuery();
+    const query = this.baseQuery()
+      .leftJoin('t.user', 'user')
+      .leftJoinAndSelect('t.category', 'category')
+      .addSelect('user.id')
+      .where('user.id = :userId', { userId: user.id });
     if (categoryName) {
       query.andWhere('t.category.name = :categoryName', { categoryName });
     }
@@ -40,6 +49,7 @@ export class TransactionService {
         { createdFrom, createdTo },
       );
     }
+
     return await query.getMany();
   }
 
@@ -50,15 +60,24 @@ export class TransactionService {
     return transaction;
   }
 
-  async update(id: number, dto: UpdateTransactionDto) {
-    const transaction = this.findOne(id);
+  async update(id: number, dto: UpdateTransactionDto, user: User) {
+    const transaction = await this.findOne(id);
+    if (transaction.user.id !== user.id)
+      throw new ForbiddenException(
+        'You are not authorized to update this transaction',
+      );
     return await this.trRepo.save({
       ...transaction,
       ...dto,
     });
   }
 
-  async remove(id: number) {
+  async remove(id: number, user: User) {
+    const transaction = await this.findOne(id);
+    if (transaction.user.id !== user.id)
+      throw new ForbiddenException(
+        'You are not authorized to delete this transaction',
+      );
     return await this.trRepo.delete(id);
   }
 }
